requirements.txt
allure-pytest==2.7.1
allure-python-commons==2.7.1
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytest-xdist==1.29.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
